Allow instack to work with pre-v2 versions of dib
We didn't branch instack in mitaka, so it's still using master.
The dib v2 changes broke mitaka because it still has an older
dib.  Rather than create a mitaka branch this late, let's just
handle both dib v1 and v2.  This only requires noop'ing the
generate_environment call when we detect an older version of
element-info.

Change-Id: Ice4be8e5f2e32733beaf8eef014f5203fce5316f
Closes-Bug: 1673424